+++ packages/components/src/switch/Switch.js
@@ -6,8 +6,10 @@ import './Switch.css';
 import { logActionRequiredIf } from '../utilities';
 
 import KeyCodes from '../common/keyCodes';
+import { useDirection } from '../common/hooks';
 
 const Switch = (props) => {
+  const { isRTL } = useDirection();
   const { checked, className, id, onClick } = props;
   const handleKeyDown = (event) => {
     if (event.code === 32 || event.keyCode === KeyCodes.SPACE) {
@@ -28,7 +30,9 @@ const Switch = (props) => {
     <span
       className={classnames(
         'np-switch',
+
         {
+          'np-switch--rtl': isRTL,
           'np-switch--unchecked': !checked,
           'np-switch--checked': checked,
         },