+++ packages/dynamic-flows/src/field/Field.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import Types from 'prop-types';
 import classNames from 'classnames';
 
-import { InlineAlert, InstructionsList, Sentiment } from '@transferwise/components';
+import { InlineAlert, InstructionsList } from '@transferwise/components';
 import FormControl from '../formControl';
 import { FormControlType, Size } from '../common';
 
@@ -101,7 +101,7 @@ export default class Field extends Component {
 
     if (this.objectSizeOf(validationFailures) > 0) {
       return {
-        type: Sentiment.ERROR,
+        type: 'error',
         content: Object.keys(validationFailures).map((key) => (
           <div key={key}>{validationFailures[key]}</div>
         )),
@@ -109,25 +109,25 @@ export default class Field extends Component {
     }
     if (error && errorMessage) {
       return {
-        type: Sentiment.ERROR,
+        type: 'error',
         content: <>{errorMessage}</>,
       };
     }
     if (warningMessage) {
       return {
-        type: Sentiment.WARNING,
+        type: 'warning',
         content: <>{warningMessage}</>,
       };
     }
     if (focused && field.help && field.help.message) {
       return {
-        type: Sentiment.INFO,
+        type: 'info',
         content: <>{field.help.message}</>,
       };
     }
     if (focused && field.help && field.help.list && field.help.list.length > 0) {
       return {
-        type: Sentiment.INFO,
+        type: 'info',
         // eslint-disable-next-line react/no-array-index-key
         content: field.help.list.map((item, index) => <div key={index}>{item}</div>),
       };
@@ -141,7 +141,7 @@ export default class Field extends Component {
       field.help.dont.length > 0
     ) {
       return {
-        type: Sentiment.INFO,
+        type: 'info',
         content: (
           <div className="m-b-1">
             <InstructionsList dos={field.help.do} donts={field.help.dont} />
@@ -151,7 +151,7 @@ export default class Field extends Component {
     }
     if (focused && control !== FormControlType.FILE && field.help && field.help.image) {
       return {
-        type: Sentiment.INFO,
+        type: 'info',
         content: <img className="thumbnail m-y-2" src={`${field.help.image}`} alt={field.label} />,
       };
     }
@@ -176,9 +176,9 @@ export default class Field extends Component {
     return (
       <div
         className={classNames('form-group', `tw-field-${name}`, {
-          'has-info': alert.type === Sentiment.INFO && showAlert,
-          'has-error': alert.type === Sentiment.ERROR && showAlert,
-          'has-warning': alert.type === Sentiment.WARNING && showAlert,
+          'has-info': alert.type === 'info' && showAlert,
+          'has-error': alert.type === 'error' && showAlert,
+          'has-warning': alert.type === 'warning' && showAlert,
           hidden: field.hidden,
         })}
       >