+++ packages/dynamic-flows/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [13.1.4](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.1.3...@transferwise/dynamic-flows@13.1.4) (2021-05-13)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
+## [13.1.3](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.1.2...@transferwise/dynamic-flows@13.1.3) (2021-05-12)
+
+
+### Bug Fixes
+
+* provider language setup ([#939](https://github.com/transferwise/neptune-web/issues/939)) ([f0c4661](https://github.com/transferwise/neptune-web/commit/f0c466150a72ab4ba6bc2f0676d0d7063f374d7e))
+
+
+
+
+
+## [13.1.2](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.1.1...@transferwise/dynamic-flows@13.1.2) (2021-05-11)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
+## [13.1.1](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.1.0...@transferwise/dynamic-flows@13.1.1) (2021-05-10)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
+# [13.1.0](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.0.3...@transferwise/dynamic-flows@13.1.0) (2021-05-10)
+
+
+### Features
+
+* Use info for help ([#932](https://github.com/transferwise/neptune-web/issues/932)) ([2644b1b](https://github.com/transferwise/neptune-web/commit/2644b1bee410fd8cce3927ca541728bcbf872efd))
+
+
+
+
+
 ## [13.0.3](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@13.0.2...@transferwise/dynamic-flows@13.0.3) (2021-05-04)
 
 **Note:** Version bump only for package @transferwise/dynamic-flows